added login layout
