+++ app/src/main/AndroidManifest.xml
@@ -14,19 +14,27 @@
         android:theme="@style/Theme.StoryCircle"
         tools:targetApi="31">
         <activity
-            android:name=".OnBoardingActivity"
+            android:name=".presentation.activity.auth.RegisterActivity"
+            android:exported="false" />
+        <activity
+            android:name=".presentation.activity.auth.LoginActivity"
             android:exported="false"
             android:theme="@style/Theme.StoryCircle.NoActionBar"/>
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.activity.OnBoardingActivity"
             android:exported="false"
-            android:theme="@style/Theme.StoryCircle.NoActionBar"/>
+            android:theme="@style/Theme.StoryCircle.NoActionBar" />
         <activity
-            android:name=".SplashActivity"
+            android:name=".presentation.activity.MainActivity"
+            android:exported="false"
+            android:theme="@style/Theme.StoryCircle.NoActionBar" />
+        <activity
+            android:name=".presentation.activity.SplashActivity"
             android:exported="true"
             android:theme="@style/Theme.StoryCircle.FullScreen">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
