+++ app/src/main/java/com/krisna/storycircle/presentation/activity/MainActivity.kt
@@ -1,10 +1,11 @@
-package com.krisna.storycircle
+package com.krisna.storycircle.presentation.activity
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.findNavController
 import androidx.navigation.ui.setupWithNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.krisna.storycircle.R
 import com.krisna.storycircle.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
