+++ app/src/main/java/com/krisna/storycircle/presentation/activity/SplashActivity.kt
@@ -1,4 +1,4 @@
-package com.krisna.storycircle
+package com.krisna.storycircle.presentation.activity
 
 import android.annotation.SuppressLint
 import android.content.Intent
