+++ app/src/main/java/com/krisna/storycircle/presentation/activity/auth/RegisterActivity.kt
@@ -1,0 +1,12 @@
+package com.krisna.storycircle.presentation.activity.auth
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.krisna.storycircle.R
+
+class RegisterActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_register)
+    }
+}
