+++ app/src/main/res/layout/fragment_add_story.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.AddStoryFragment">
+    tools:context=".presentation.fragment.AddStoryFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
