+++ app/src/main/res/navigation/bot_nav_navigation.xml
@@ -7,17 +7,17 @@
 
     <fragment
         android:id="@+id/menu_home"
-        android:name="com.krisna.storycircle.fragment.HomeFragment"
+        android:name="com.krisna.storycircle.presentation.fragment.HomeFragment"
         android:label="Home"
         tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/menu_addstory"
-        android:name="com.krisna.storycircle.fragment.AddStoryFragment"
+        android:name="com.krisna.storycircle.presentation.fragment.AddStoryFragment"
         android:label="Add Story"
         tools:layout="@layout/fragment_add_story" />
     <fragment
         android:id="@+id/menu_settings"
-        android:name="com.krisna.storycircle.fragment.SettingsFragment"
+        android:name="com.krisna.storycircle.presentation.fragment.SettingsFragment"
         android:label="Settings"
         tools:layout="@layout/fragment_settings" />
