+++ app/src/main/res/values/strings.xml
@@ -2,6 +2,11 @@
     <string name="app_name">Story Circle</string>
     <string name="onboarding_message">Join our community and start sharing your stories today!</string>
     <string name="btn_get_started">Get Started</string>
-    <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="email_hint">Email</string>
+    <string name="password_hint">Password</string>
+    <string name="login_description">Welcome back, you\'ve been missed!</string>
+    <string name="login_title">Login</string>
+    <string name="donthaveacc_description">Don\'t have an account?</string>
+    <string name="register">Register</string>
 </resources>
